Add App global Fixture
